+++ src/FakeXrmEasy.Plugins/IXrmFakedContextPipelineExtensions.cs
@@ -19,6 +19,7 @@ using FakeXrmEasy.Plugins.PluginSteps.PluginStepRegistrationFieldNames;
 using FakeXrmEasy.Plugins.PluginSteps.Extensions;
 using FakeXrmEasy.Plugins.Definitions;
 using FakeXrmEasy.Plugins.Pipeline.PipelineTypes;
+using FakeXrmEasy.Plugins.PluginInstances;
 
 namespace FakeXrmEasy.Pipeline
 {
@@ -111,13 +112,6 @@ namespace FakeXrmEasy.Pipeline
             var entityType = entity.GetType();
             if (entityType.IsSubclassOf(typeof(Entity)))
             {
-                var entityTypeCodeField = entityType.GetField("EntityTypeCode");
-                if (entityTypeCodeField == null)
-                {
-                    throw new EntityTypeCodeNotFoundException(entity.LogicalName);
-                }
-                var entityTypeCode = (int)entityTypeCodeField.GetValue(entity);
-
                 var pluginStepDefinition = new PluginStepDefinition()
                 {
                     MessageName = message,
@@ -125,7 +119,6 @@ namespace FakeXrmEasy.Pipeline
                     Mode = mode,
                     Rank = rank,
                     FilteringAttributes = filteringAttributes,
-                    EntityTypeCode = entityTypeCode,
                     EntityLogicalName = entity.LogicalName,
                     ImagesDefinitions = registeredImages
                 };
@@ -232,17 +225,20 @@ namespace FakeXrmEasy.Pipeline
                 var hasSdkMessageFilter = context.FindReflectedType(PluginStepRegistrationEntityNames.SdkMessageFilter) != null;
                 var hasSdkMessageProcessingStep = context.FindReflectedType(PluginStepRegistrationEntityNames.SdkMessageProcessingStep) != null;
                 var hasPluginType = context.FindReflectedType(PluginStepRegistrationEntityNames.PluginType) != null;
-
+                var hasSdkMessageProcessingStepSecureConfig = context.FindReflectedType(PluginStepRegistrationEntityNames.SdkMessageProcessingStepSecureConfig) != null;
+                
                 var hasAllPipelineTypes =
                     hasSdkMessage &&
                     hasSdkMessageFilter &&
                     hasSdkMessageProcessingStep &&
+                    hasSdkMessageProcessingStepSecureConfig &&
                     hasPluginType;
 
                 var hasSomePipelineTypes =
                     hasSdkMessage ||
                     hasSdkMessageFilter ||
                     hasSdkMessageProcessingStep ||
+                    hasSdkMessageProcessingStepSecureConfig ||
                     hasPluginType;
 
                 if(hasAllPipelineTypes)
@@ -296,6 +292,26 @@ namespace FakeXrmEasy.Pipeline
             }
             return sdkFilter;
         }
+        
+        internal static Entity AddSdkMessageProcessingStepSecureConfig(IXrmFakedContext context,
+            IPluginStepDefinition pluginStepDefinition)
+        {
+            Entity sdkMessageProcessingStepSecureConfig = null;
+            if (pluginStepDefinition.Configurations != null)
+            {
+                var sdkMessageProcessingSecureConfigId = pluginStepDefinition.Configurations.SecureConfigId;
+                sdkMessageProcessingStepSecureConfig = new Entity(PluginStepRegistrationEntityNames.SdkMessageProcessingStepSecureConfig)
+                {
+                    Id = sdkMessageProcessingSecureConfigId,
+                    [SdkMessageProcessingStepSecureConfigFieldNames.SecureConfig] =
+                        pluginStepDefinition.Configurations.SecureConfig
+                };
+                
+                context.AddEntityWithDefaults(sdkMessageProcessingStepSecureConfig);
+            }
+
+            return sdkMessageProcessingStepSecureConfig;
+        }
 
         internal static void AddSdkMessageProcessingStepImages(IXrmFakedContext context,
                                                 IPluginStepDefinition pluginStepDefinition,
@@ -327,11 +343,41 @@ namespace FakeXrmEasy.Pipeline
             return context.RegisterPluginStepInternal(typeof(TPlugin), pluginStepDefinition);
         }
 
+        internal static void VerifyPluginStepDefinition(IXrmFakedContext context,
+            IPluginStepDefinition pluginStepDefinition)
+        {
+            if (pluginStepDefinition.EntityTypeCode != null &&
+                !string.IsNullOrWhiteSpace(pluginStepDefinition.EntityLogicalName))
+            {
+                throw new RegisterStepWithEntityTypeCodeAndEntityLogicalNameException(
+                    pluginStepDefinition.EntityTypeCode.Value, pluginStepDefinition.EntityLogicalName);
+            }
+
+            if (pluginStepDefinition.EntityTypeCode != null)
+            {
+                if (!context.ProxyTypesAssemblies.Any())
+                {
+                    throw new RegisterEntityTypeCodePluginStepWithoutProxyTypesAssemblyException(pluginStepDefinition
+                        .EntityTypeCode.Value);
+                }
+                
+                var reflectedType = context.FindReflectedType(pluginStepDefinition.EntityTypeCode.Value);
+                if (reflectedType == null)
+                {
+                    throw new RegisterInvalidEntityTypeCodePluginStepException(pluginStepDefinition
+                        .EntityTypeCode.Value);
+                }
+                
+                pluginStepDefinition.EntityLogicalName = (string)reflectedType.GetField("EntityLogicalName")?.GetValue(null);
+            }
+        }
+        
         internal static Guid RegisterPluginStepInternal(this IXrmFakedContext context,
                                                                 Type pluginType,
                                                                 IPluginStepDefinition pluginStepDefinition)
 
         {
+            VerifyPluginStepDefinition(context, pluginStepDefinition);
             ValidatePluginStep(context, pluginStepDefinition);
 
             AddPipelineTypes(context);
@@ -344,6 +390,10 @@ namespace FakeXrmEasy.Pipeline
             var assemblyName = pluginType.Assembly.GetName();
             var pluginTypeRecord = context.AddPluginType(pluginType, assemblyName);
 
+            // Secure config
+            var sdkMessageProcessingStepSecureConfig =
+                AddSdkMessageProcessingStepSecureConfig(context, pluginStepDefinition);
+            
             // Message Step
             var sdkMessageProcessingStepId = pluginStepDefinition.Id == Guid.Empty ? Guid.NewGuid() : pluginStepDefinition.Id;
 
@@ -356,15 +406,24 @@ namespace FakeXrmEasy.Pipeline
                 [SdkMessageProcessingStepFieldNames.FilteringAttributes] = pluginStepDefinition.FilteringAttributes != null ? string.Join(",", pluginStepDefinition.FilteringAttributes) : null,
                 [SdkMessageProcessingStepFieldNames.Stage] = new OptionSetValue((int)pluginStepDefinition.Stage),
                 [SdkMessageProcessingStepFieldNames.Mode] = new OptionSetValue((int)pluginStepDefinition.Mode),
-                [SdkMessageProcessingStepFieldNames.Rank] = pluginStepDefinition.Rank
+                [SdkMessageProcessingStepFieldNames.Rank] = pluginStepDefinition.Rank,
+                [SdkMessageProcessingStepFieldNames.Configuration] = pluginStepDefinition.Configurations?.UnsecureConfig,
+                [SdkMessageProcessingStepFieldNames.SdkMessageProcessingStepSecureConfigId] = sdkMessageProcessingStepSecureConfig?.ToEntityReference()
             };
             context.AddEntityWithDefaults(sdkMessageProcessingStep);
 
+            if (pluginStepDefinition.PluginInstance != null)
+            {
+                var pluginInstancesRepository = context.GetProperty<IPluginInstancesRepository>();
+                pluginInstancesRepository.SetPluginInstance(sdkMessageProcessingStepId, pluginStepDefinition.PluginInstance);
+            }
+            
             AddSdkMessageProcessingStepImages(context, pluginStepDefinition, sdkMessageProcessingStep);
 
             return sdkMessageProcessingStepId;
         }
 
+        [Obsolete("This method is obsolete, should no longer be used as it has been replaced by a more efficient method 'GetPluginStepsForOrganizationRequest' that instead directly queries the FakeXrmEasy internal in-memory database")]
         internal static IEnumerable<PluginStepDefinition> GetPluginStepsForOrganizationRequestWithRetrieveMultiple(this IXrmFakedContext context, string requestName, ProcessingStepStage stage, ProcessingStepMode mode, OrganizationRequest request)
         {
             var target = GetTargetForRequest(request);
@@ -445,8 +504,31 @@ namespace FakeXrmEasy.Pipeline
             }
         }
 
+        private static void ExecutePlugin(IXrmFakedContext context,
+            PluginStepDefinition pluginStepDefinition,
+            XrmFakedPluginExecutionContext pluginContext,
+            bool isAuditEnabled)
+        {
+            var pluginMethod = GetPluginMethod(pluginStepDefinition);
+            
+            try
+            {
+                InvokePluginMethod(pluginMethod, context, pluginStepDefinition, pluginContext);
+            }
+            catch (TargetInvocationException ex)
+            {
+                throw ex.InnerException;
+            }
+            
+            if (isAuditEnabled)
+            {
+                context.AddPluginStepAuditDetails(pluginMethod, pluginContext, pluginStepDefinition);
+            }
+        }
+            
+            
         /// <summary>
-        /// 
+        /// Executes all the relevant plugin steps for the current request
         /// </summary>
         /// <param name="context"></param>
         /// <param name="pluginSteps"></param>
@@ -465,8 +547,6 @@ namespace FakeXrmEasy.Pipeline
 
             foreach (var pluginStep in pluginSteps)
             {
-                var pluginMethod = GetPluginMethod(pluginStep);
-
                 IEnumerable<Entity> preImageDefinitions = null;
                 if (previousValues != null)
                 {
@@ -488,19 +568,7 @@ namespace FakeXrmEasy.Pipeline
                 pluginContext.PreEntityImages = GetEntityImageCollection(preImageDefinitions, previousValues);
                 pluginContext.PostEntityImages = GetEntityImageCollection(postImageDefinitions, resultingAttributes);
 
-                try
-                {
-                    pluginMethod.Invoke(null, new object[] { context, pluginContext });
-                }
-                catch (TargetInvocationException ex)
-                {
-                    throw ex.InnerException;
-                }
-
-                if (isAuditEnabled)
-                {
-                    context.AddPluginStepAuditDetails(pluginMethod, pluginContext, pluginStep);
-                }
+                ExecutePlugin(context, pluginStep, pluginContext, isAuditEnabled);
             }
         }
 
@@ -517,7 +585,8 @@ namespace FakeXrmEasy.Pipeline
                 MessageName = pluginContext.MessageName,
                 Stage = (ProcessingStepStage)pluginContext.Stage,
                 InputParameters = pluginContext.InputParameters,
-                OutputParameters = pluginContext.OutputParameters
+                OutputParameters = pluginContext.OutputParameters,
+                PluginStepDefinition = pluginStep
             };
 
             if (pluginContext.InputParameters.ContainsKey("Target"))
@@ -530,8 +599,6 @@ namespace FakeXrmEasy.Pipeline
                     pluginStepAuditDetails.TargetEntityReference = (EntityReference)target;
             }
 
-
-
             var pluginStepAudit = context.GetProperty<IPluginStepAudit>() as PluginStepAudit;
             pluginStepAudit.Add(pluginStepAuditDetails);
         }
@@ -541,10 +608,36 @@ namespace FakeXrmEasy.Pipeline
             var assembly = AppDomain.CurrentDomain.Load(pluginStepDefinition.AssemblyName);
             var pluginType = assembly.GetType(pluginStepDefinition.PluginType);
 
-            var methodInfo = typeof(IXrmBaseContextPluginExtensions).GetMethod("ExecutePluginWith", new[] { typeof(IXrmFakedContext), typeof(XrmFakedPluginExecutionContext) });
-            var pluginMethod = methodInfo.MakeGenericMethod(pluginType);
-
-            return pluginMethod;
+            MethodInfo methodInfo = null;
+            if (pluginStepDefinition.Configurations != null)
+            {
+                methodInfo = typeof(IXrmBaseContextPluginExtensions).GetMethod("ExecutePluginWithConfigurations", new[] { typeof(IXrmFakedContext), typeof(XrmFakedPluginExecutionContext), typeof(string), typeof(string) });
+            }
+            else
+            {
+                methodInfo = typeof(IXrmBaseContextPluginExtensions).GetMethod("ExecutePluginWith", new[] { typeof(IXrmFakedContext), typeof(XrmFakedPluginExecutionContext) });
+            }
+            
+            return methodInfo.MakeGenericMethod(pluginType);
+        }
+        
+        private static void InvokePluginMethod(MethodInfo pluginMethod, 
+            IXrmFakedContext context,
+            IPluginStepDefinition pluginStepDefinition, 
+            XrmFakedPluginExecutionContext pluginContext)
+        {
+            if (pluginStepDefinition.Configurations != null)
+            {
+                pluginMethod.Invoke(null, new object[] { context, 
+                    pluginContext, 
+                    pluginStepDefinition.Configurations.UnsecureConfig,
+                    pluginStepDefinition.Configurations.SecureConfig
+                });
+            }
+            else
+            {
+                pluginMethod.Invoke(null, new object[] { context, pluginContext });
+            }
         }
 
         private static IEnumerable<PluginStepDefinition> GetStepsForStageWithRetrieveMultiple(this IXrmFakedContext context, string requestName, ProcessingStepStage stage, ProcessingStepMode mode, Entity entity)
@@ -654,12 +747,17 @@ namespace FakeXrmEasy.Pipeline
                 entityLogicalName = entity.LogicalName;
             }
 
+            var pluginInstancesRepository = context.GetProperty<IPluginInstancesRepository>();
+            
             var pluginSteps = (from step in context.CreateQuery(PluginStepRegistrationEntityNames.SdkMessageProcessingStep)
                                join message in context.CreateQuery(PluginStepRegistrationEntityNames.SdkMessage) on (step[SdkMessageProcessingStepFieldNames.SdkMessageId] as EntityReference).Id equals message.Id
                                join pluginAssembly in context.CreateQuery(PluginStepRegistrationEntityNames.PluginType) on (step[SdkMessageProcessingStepFieldNames.EventHandler] as EntityReference).Id equals pluginAssembly.Id
                                join mFilter in context.CreateQuery(PluginStepRegistrationEntityNames.SdkMessageFilter) on (step[SdkMessageProcessingStepFieldNames.SdkMessageFilterId] != null ? (step[SdkMessageProcessingStepFieldNames.SdkMessageFilterId] as EntityReference).Id : Guid.Empty) equals mFilter.Id into mesFilter
-                               from messageFilter in mesFilter.DefaultIfEmpty()
+                               join stepSecConfig in context.CreateQuery(PluginStepRegistrationEntityNames.SdkMessageProcessingStepSecureConfig) on (step[SdkMessageProcessingStepFieldNames.SdkMessageProcessingStepSecureConfigId] != null ? (step[SdkMessageProcessingStepFieldNames.SdkMessageProcessingStepSecureConfigId] as EntityReference).Id : Guid.Empty) equals stepSecConfig.Id into secureConfig
 
+                               from messageFilter in mesFilter.DefaultIfEmpty()
+                               from secureConfiguration in secureConfig.DefaultIfEmpty()
+                               
                                where (step[SdkMessageProcessingStepFieldNames.Stage] as OptionSetValue).Value == (int)stage
                                where (step[SdkMessageProcessingStepFieldNames.Mode] as OptionSetValue).Value == (int)mode
                                where (message[SdkMessageFieldNames.Name] as string) == requestName
@@ -674,7 +772,14 @@ namespace FakeXrmEasy.Pipeline
                                    EntityTypeCode = messageFilter.GetMessageFilterPrimaryObjectCode(),
                                    EntityLogicalName = messageFilter.GetMessageFilterEntityLogicalName(),
                                    AssemblyName = pluginAssembly.GetAttributeValue<string>(PluginTypeFieldNames.AssemblyName),
-                                   PluginType = pluginAssembly.GetAttributeValue<string>(PluginTypeFieldNames.TypeName)
+                                   PluginType = pluginAssembly.GetAttributeValue<string>(PluginTypeFieldNames.TypeName),
+                                   Configurations = secureConfiguration != null ? new PluginStepConfigurations()
+                                   {
+                                       SecureConfigId = secureConfiguration.Id,
+                                       SecureConfig = secureConfiguration.GetAttributeValue<string>(SdkMessageProcessingStepSecureConfigFieldNames.SecureConfig),
+                                       UnsecureConfig = step.GetAttributeValue<string>(SdkMessageProcessingStepFieldNames.Configuration)
+                                   } : null,
+                                   PluginInstance = pluginInstancesRepository.GetPluginInstance(step.Id)
                                }).OrderBy(ps => ps.Rank)
                                  .ToList();